tests/wpt/tests/css/css-overflow/scrollable-overflow-with-grid-item-margin-inline-end.html

<!DOCTYPE html>
<title>CSS Overflow: Scrollable overflow from grid item with margin-inline-end</title>
<link rel="help" href="https://drafts.csswg.org/css-overflow-3/#scrollable" />
<meta name="assert" content="Grid item contribute its margin-end to parent scroller's scrollable overflow.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  #container {
    width: 100px;
    height: 100px;
    display: grid;
  }

  .item {
    width: 50px;
    height: 50px;
    margin-inline-end: 950px;
  }
</style>

<div id=container>
  <div class=item></div>
</div>
<script>
  test(() => {
    container.style.overflow = "auto";
    assert_equals(container.scrollWidth, 1000);
  }, "Check scrollWidth with overflow: auto");

  test(() => {
    container.style.overflow = "scroll";
    assert_equals(container.scrollWidth, 1000);
  }, "Check scrollWidth with overflow: scroll");

  test(() => {
    container.style.overflow = "hidden";
    assert_equals(container.scrollWidth, 1000);
  }, "Check scrollWidth with overflow: hidden");

  test(() => {
    container.style.overflow = "clip";
    assert_equals(container.scrollWidth, 100);
  }, "Check scrollWidth with overflow: clip");

  test(() => {
    container.style.overflow = "visible";
    assert_equals(container.scrollWidth, 100);
  }, "Check scrollWidth with overflow: visible");

  test(() => {
    container.style.overflowX = "visible";
    container.style.overflowY = "clip";
    assert_equals(container.scrollWidth, 100);
  }, "Check scrollWidth with overflowX: visible, overflowY: clip");

  test(() => {
    container.style.overflowX = "visible";
    container.style.overflowY = "hidden";
    assert_equals(container.scrollWidth, 1000);
  }, "Check scrollWidth with overflowX: visible, overflowY: hidden");

  test(() => {
    container.style.overflowX = "visible";
    container.style.overflowY = "auto";
    assert_equals(container.scrollWidth, 1000);
  }, "Check scrollWidth with overflowX: visible, overflowY: auto");
</script>
